+++ crates/zng-ext-l10n/src/types.rs
@@ -269,18 +269,22 @@ pub struct L10nSpecialize<T>(pub Option<T>);
 pub trait IntoL10nVar {
     fn to_l10n_var(&mut self) -> Var<L10nArgument>;
 }
-
-impl<T: Into<L10nArgument>> IntoL10nVar for L10nSpecialize<T> {
+impl<T: fmt::Display> IntoL10nVar for L10nSpecialize<T> {
+    fn to_l10n_var(&mut self) -> Var<L10nArgument> {
+        const_var(self.0.take().unwrap().to_txt().into())
+    }
+}
+impl<T: Into<L10nArgument>> IntoL10nVar for &mut L10nSpecialize<T> {
     fn to_l10n_var(&mut self) -> Var<L10nArgument> {
         const_var(self.0.take().unwrap().into())
     }
 }
-impl<T: VarValue + Into<L10nArgument>> IntoL10nVar for &mut L10nSpecialize<Var<T>> {
+impl<T: VarValue + Into<L10nArgument>> IntoL10nVar for &mut &mut L10nSpecialize<Var<T>> {
     fn to_l10n_var(&mut self) -> Var<L10nArgument> {
         self.0.take().unwrap().map_into()
     }
 }
-impl IntoL10nVar for &mut &mut L10nSpecialize<Var<L10nArgument>> {
+impl IntoL10nVar for &mut &mut &mut L10nSpecialize<Var<L10nArgument>> {
     fn to_l10n_var(&mut self) -> Var<L10nArgument> {
         self.0.take().unwrap()
     }